Refactoring vuex store
