+++ Section16 - Find a Coach/.gitignore
@@ -1,6 +1,7 @@
 .DS_Store
 node_modules
 /dist
+.firebase
 
 # local env files
 .env.local
