+++ Section16 - Find a Coach/src/store/index.ts
@@ -3,20 +3,13 @@ import { createStore } from 'vuex';
 import { coachesModule } from './modules/coaches';
 import { requestsModule } from './modules/requests';
 import { authModule } from './modules/auth';
-import { RootState } from '@/types/interfaces';
+import { RootState } from '@/store/rootTypes';
 
-const state: RootState = {
-  root: true,
-  version: '1.0',
-};
-
-export const store = createStore({
-  state,
+export default createStore<RootState>({
+  state: {},
   modules: {
     auth: authModule,
     coaches: coachesModule,
     requests: requestsModule,
   },
 });
-
-export default store;
