--- Section16 - Find a Coach/src/store/modules/auth/action-types.ts
@@ -1,8 +1,0 @@
-export enum AuthActionTypes {
-  login = 'login',
-  signUp = 'signUp',
-  auth = 'auth',
-  tryLogin = 'tryLogin',
-  logout = 'logout',
-  autoLogout = 'autoLogout',
-}
