--- Section16 - Find a Coach/src/store/modules/auth/mutation-types.ts
@@ -1,4 +1,0 @@
-export enum AuthMutationTypes {
-  setUser = 'setUser',
-  setAutoLogout = 'setAutoLogout',
-}
