--- Section16 - Find a Coach/src/store/modules/auth/state.ts
@@ -1,7 +1,0 @@
-import { AuthModuleState } from '@/types/interfaces';
-
-export const state: AuthModuleState = {
-  userId: null,
-  token: null,
-  didAutoLogout: false,
-};
