--- Section16 - Find a Coach/src/store/modules/coaches/action-types.ts
@@ -1,4 +1,0 @@
-export enum CoachesActionTypes {
-  registerCoach = 'registerCoach',
-  loadCoaches = 'loadCoaches',
-}
