--- Section16 - Find a Coach/src/store/modules/coaches/mutation-types.ts
@@ -1,5 +1,0 @@
-export enum CoachesMutationTypes {
-  registerCoach = 'registerCoach',
-  setCoaches = 'setCoaches',
-  setFetchTimestamp = 'setFetchTimestamp',
-}
