--- Section16 - Find a Coach/src/store/modules/requests/action-types.ts
@@ -1,4 +1,0 @@
-export enum RequestActionTypes {
-  contactCoach = 'contactCoach',
-  fetchRequests = 'fetchRequests',
-}
