--- Section16 - Find a Coach/src/store/modules/requests/mutation-types.ts
@@ -1,4 +1,0 @@
-export enum RequestsMutationTypes {
-  addRequest = 'addRequest',
-  setRequests = 'setRequests',
-}
