--- Section16 - Find a Coach/src/store/modules/requests/state.ts
@@ -1,5 +1,0 @@
-import { RequestsModuleState } from '@/types/interfaces';
-
-export const state: RequestsModuleState = {
-  requests: [],
-};
