+++ Section16 - Find a Coach/src/store/rootTypes.ts
@@ -1,0 +1,1 @@
+export type RootState = {};
